+++ lib/DocblockParser/Parser.php
@@ -49,6 +49,7 @@ use Phpactor\DocblockParser\Ast\Tag\ThrowsTag;
 use Phpactor\DocblockParser\Ast\VariableNode;
 use Phpactor\DocblockParser\Ast\Token;
 use Phpactor\DocblockParser\Ast\Tokens;
+use RuntimeException;
 
 final class Parser
 {
@@ -88,7 +89,7 @@ final class Parser
 
     private function parseTag(): TagNode
     {
-        $token = $this->tokens->current;
+        $token = $this->tokens->mustGetCurrent();
         $value = str_replace(['@psalm-', '@phpstan-'], '@', $token->value);
         return match ($value) {
             '@param' => $this->parseParam(),
@@ -103,14 +104,14 @@ final class Parser
             '@template-covariant' => $this->parseTemplate(),
             '@extends', '@template-extends' => $this->parseExtends(),
             '@implements', '@template-implements' => $this->parseImplements(),
-            default => new UnknownTag($this->tokens->chomp()),
+            default => new UnknownTag($this->tokens->mustChomp()),
         };
     }
 
     private function parseParam(): ParamTag
     {
         $type = $variable = $textNode = null;
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
 
         if ($this->ifType()) {
             $type = $this->parseTypes();
@@ -120,12 +121,12 @@ final class Parser
             $variable = $this->parseVariable();
         }
 
-        return new ParamTag($tag, $type, $variable, $this->parseText());
+        return new ParamTag($tag, $type, $variable, $this->parseText() ?: new TextNode([]));
     }
 
     private function parseVar(): VarTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = $variable = null;
         if ($this->ifType()) {
             $type = $this->parseTypes();
@@ -139,7 +140,7 @@ final class Parser
 
     private function parseThrows(): ThrowsTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = null;
 
         if ($this->tokens->if(Token::T_LABEL)) {
@@ -151,12 +152,12 @@ final class Parser
 
     private function parseMethod(): MethodTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = $name = $parameterList = $open = $close = null;
         $static = null;
 
         if ($this->tokens->ifNextIs(Token::T_LABEL)) {
-            if ($this->tokens->current->value === 'static') {
+            if ($this->tokens->mustGetCurrent()->value === 'static') {
                 $static = $this->tokens->chomp();
             }
         }
@@ -180,7 +181,7 @@ final class Parser
 
     private function parseProperty(): PropertyTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = $name = null;
         if ($this->ifType()) {
             $type = $this->parseTypes();
@@ -206,7 +207,7 @@ final class Parser
                 $this->tokens->if(Token::T_BAR) ||
                 $this->tokens->if(Token::T_AMPERSAND)
             ) {
-                $delimiter = $this->tokens->chomp();
+                $delimiter = $this->tokens->mustChomp();
                 if (!$mode) {
                     $mode = $delimiter->type;
                 }
@@ -245,19 +246,23 @@ final class Parser
 
         if ($this->tokens->current->type === Token::T_VARIABLE) {
             if ($this->tokens->current->value === '$this') {
-                $variable = $this->tokens->chomp(Token::T_VARIABLE);
+                $variable = $this->tokens->mustChomp(Token::T_VARIABLE);
                 return new ThisNode($variable);
             }
             return $this->parseConditionalType();
         }
 
         if ($this->tokens->current->type === Token::T_NULLABLE) {
-            $nullable = $this->tokens->chomp();
-            return new NullableNode($nullable, $this->parseTypes());
+            $nullable = $this->tokens->mustChomp();
+            $type = $this->parseType();
+            if ($type === null) {
+                return null;
+            }
+            return new NullableNode($nullable, $type);
         }
 
         if ($this->tokens->current->type === Token::T_PAREN_OPEN) {
-            $open = $this->tokens->chomp();
+            $open = $this->tokens->mustChomp();
             $this->tokens->chompWhitespace();
             $type = $this->parseTypes();
             $this->tokens->chompWhitespace();
@@ -266,9 +271,10 @@ final class Parser
             return new ParenthesizedType($open, $type, $close);
         }
 
-        $type = $this->tokens->chomp();
+        $type = $this->tokens->mustChomp();
 
-        if (null === $this->tokens->current) {
+        /** @phpstan-ignore-next-line It can be null*/
+        if (null === $this->tokens->current && $type) {
             return $this->createTypeFromToken($type);
         }
 
@@ -304,7 +310,7 @@ final class Parser
         }
 
         if ($this->tokens->current->type === Token::T_BRACKET_ANGLE_OPEN) {
-            $open = $this->tokens->chomp();
+            $open = $this->tokens->mustChomp();
             $typeList = null;
             if ($this->tokens->if(Token::T_VARIABLE)) {
                 $typeList = $this->parseTypeList();
@@ -316,6 +322,9 @@ final class Parser
                 $typeList = $this->parseTypeList();
             }
             if ($this->tokens->if(Token::T_INTEGER)) {
+                $typeList = $this->parseTypeList();
+            }
+            if ($this->tokens->if(Token::T_NULLABLE)) {
                 $typeList = $this->parseTypeList();
             }
 
@@ -331,7 +340,7 @@ final class Parser
                 $open,
                 $this->createTypeFromToken($type),
                 $typeList,
-                $this->tokens->chomp()
+                $this->tokens->mustChomp()
             );
             return $this->parseDimensions($type);
         }
@@ -361,7 +370,7 @@ final class Parser
     private function parseDimensions(TypeNode $type): TypeNode
     {
         while ($this->tokens->if(Token::T_LIST)) {
-            $list = $this->tokens->chomp();
+            $list = $this->tokens->mustChomp();
             $type = new ListBracketsNode($type, $list);
         }
 
@@ -403,8 +412,8 @@ final class Parser
         ) {
             return new ConstantNode(
                 $classNode,
-                $this->tokens->chomp(),
-                $this->tokens->chomp(),
+                $this->tokens->mustChomp(),
+                $this->tokens->mustChomp(),
             );
         }
 
@@ -413,11 +422,11 @@ final class Parser
 
     private function parseVariable(): ?VariableNode
     {
-        if ($this->tokens->current->type !== Token::T_VARIABLE) {
+        if ($this->tokens->mustGetCurrent()->type !== Token::T_VARIABLE) {
             return null;
         }
 
-        $name = $this->tokens->chomp(Token::T_VARIABLE);
+        $name = $this->tokens->mustChomp(Token::T_VARIABLE);
 
         return new VariableNode($name);
     }
@@ -428,6 +437,8 @@ final class Parser
         while (true) {
             if ($this->tokens->if(Token::T_LABEL)) {
                 $types[] = $this->parseTypes();
+            } elseif ($this->tokens->if(Token::T_NULLABLE)) {
+                $types[] = $this->parseTypes();
             } elseif ($this->tokens->if(Token::T_QUOTED_STRING)) {
                 $types[] = $this->parseTypes();
             } elseif ($this->tokens->if(Token::T_INTEGER)) {
@@ -436,7 +447,7 @@ final class Parser
                 $types[] = $this->parseTypes();
             }
             if ($this->tokens->if(Token::T_COMMA)) {
-                $types[] = $this->tokens->chomp();
+                $types[] = $this->tokens->mustChomp();
                 continue;
             }
             break;
@@ -455,7 +466,7 @@ final class Parser
         while (true) {
             $parameters[] = $this->parseParameter();
             if ($this->tokens->if(Token::T_COMMA)) {
-                $parameters[] = $this->tokens->chomp();
+                $parameters[] = $this->tokens->mustChomp();
                 continue;
             }
             break;
@@ -483,14 +494,14 @@ final class Parser
     private function parseDeprecated(): DeprecatedTag
     {
         return new DeprecatedTag(
-            $this->tokens->chomp(Token::T_TAG),
+            $this->tokens->mustChomp(Token::T_TAG),
             $this->parseText()
         );
     }
 
     private function parseMixin(): MixinTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = null;
 
         if ($this->tokens->if(Token::T_LABEL)) {
@@ -505,7 +516,7 @@ final class Parser
 
     private function parseReturn(): ReturnTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = null;
 
         if ($this->ifType()) {
@@ -513,7 +524,7 @@ final class Parser
         }
 
         if ($this->tokens->if(Token::T_VARIABLE)) {
-            $variable = $this->tokens->chomp(Token::T_VARIABLE);
+            $variable = $this->tokens->mustChomp(Token::T_VARIABLE);
             if ($variable->value === '$this') {
                 $type = new ThisNode($variable);
             }
@@ -542,7 +553,7 @@ final class Parser
             if ($this->tokens->current->type === Token::T_TAG) {
                 break;
             }
-            $text[] = $this->tokens->chomp();
+            $text[] = $this->tokens->mustChomp();
         }
 
         if ($text) {
@@ -565,8 +576,8 @@ final class Parser
     private function parseValue(): ?ValueNode
     {
         if ($this->tokens->if(Token::T_LABEL)) {
-            if (strtolower($this->tokens->current->value) === 'null') {
-                return new NullValue($this->tokens->chomp());
+            if (strtolower($this->tokens->mustGetCurrent()->value) === 'null') {
+                return new NullValue($this->tokens->mustChomp());
             }
         }
 
@@ -575,17 +586,17 @@ final class Parser
 
     private function parseTemplate(): TemplateTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $placeholder = null;
         $of = null;
         $type = null;
 
         if ($this->tokens->if(Token::T_LABEL)) {
-            $placeholder = $this->tokens->chomp();
+            $placeholder = $this->tokens->mustChomp();
         }
 
         if ($this->tokens->if(Token::T_LABEL)) {
-            $of = $this->tokens->chomp();
+            $of = $this->tokens->mustChomp();
             if ($of->value === 'of') {
                 /** @phpstan-ignore-next-line */
                 if ($this->tokens->if(Token::T_LABEL)) {
@@ -601,7 +612,7 @@ final class Parser
 
     private function parseExtends(): ExtendsTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $type = null;
 
         if ($this->tokens->if(Token::T_LABEL)) {
@@ -613,7 +624,7 @@ final class Parser
 
     private function parseImplements(): ImplementsTag
     {
-        $tag = $this->tokens->chomp(Token::T_TAG);
+        $tag = $this->tokens->mustChomp(Token::T_TAG);
         $types = [];
 
         if ($this->tokens->if(Token::T_LABEL)) {
@@ -685,10 +696,13 @@ final class Parser
     private function parseConditionalType(): TypeNode
     {
         $variable = $this->parseVariable();
+        if (!$variable) {
+            throw new RuntimeException('Expected a variable, this should not happen');
+        }
         if (!$this->tokens->if(Token::T_LABEL)) {
             return new ConditionalNode($variable);
         }
-        $is = $this->tokens->chomp();
+        $is = $this->tokens->mustChomp();
         if ($is->toString() !== 'is') {
             return new ConditionalNode($variable, $is);
         }