+++ lib/Extension/LanguageServerCompletion/Handler/CompletionHandler.php
@@ -178,6 +178,7 @@ class CompletionHandler implements Handler, CanRegisterCapabilities
             ':',
             '>',
             '$',
+            '[',
             '@',
             '(',
             '\'',