+++ lib/Phpactor.php
@@ -34,6 +34,8 @@ use Phpactor\Extension\LanguageServer\LanguageServerExtension;
 use Phpactor\Extension\LanguageServer\LanguageServerExtraExtension;
 use Phpactor\Extension\Laravel\LaravelExtension;
 use Phpactor\Extension\ObjectRenderer\ObjectRendererExtension;
+use Phpactor\Extension\PhpCodeSniffer\PhpCodeSnifferExtension;
+use Phpactor\Extension\PhpCodeSniffer\PhpCodeSnifferSuggestExtension;
 use Phpactor\Extension\PHPUnit\PHPUnitExtension;
 use Phpactor\Extension\Prophecy\ProphecyExtension;
 use Phpactor\Extension\Prophecy\ProphecySuggestExtension;
@@ -177,6 +179,8 @@ class Phpactor
             LanguageServerPhpCsFixerExtension::class,
 
             LanguageServerPhpCsFixerSuggestExtension::class,
+            PhpCodeSnifferExtension::class,
+            PhpCodeSnifferSuggestExtension::class,
 
             LanguageServerBlackfireExtension::class,